Add cmd with convert command
We want to be able to run the GOBL to XRechnung conversion
from the command line. To do that we add a simple command line
implementation to the cmd/gobl.xinvoice and add some logic
for the commands to work. At the moment the command generates
only a simple xml file that still needs to be populated.

Co-authored-by: Grzegorz Lisowski <noplisy@users.noreply.github.com>
Co-authored-by: Alex Malaszkiewicz <torrocus@users.noreply.github.com>