Create XRechnung validation workflow
I implemented workflow for XRechnung validation using GitHub Actions.
Launching the GitHub Actions workflow requires creating the workflow
in the main branch; otherwise, it cannot be triggered.
Consequently, I had to initiate the work in a separate repository.

I opted for triggering it using workflow dispatch since
we will have multiple XML files,
some intentionally valid, and some not.
Therefore, we need the ability to identify issues for testing purposes.
Running it on CI allows us to input the XML file path
into the input field.
This should significantly facilitate testing work
without the need to install the validator locally.

The source repository on which I tested the workflow: [1]
On the action list, [2]
you can see those that have succeeded (triggered on invoice valid)
and those that have failed (triggered on invoice invalid).

In the case of failure, under "Summary" in the "Artifacts" section,
you can find an archive containing reports.

[1]: https://github.com/torrocus/xrechnung-validator-workflow
[2]: https://github.com/torrocus/xrechnung-validator-workflow/actions