Handle correction invoices
We want to be able correct wrong invoices. To do that we
need to be able to create correction invoices. To create
such a invoice we need to set the correct invoice type and
add information about the preceding invoice.

Here we add InvoiceReferencedDocument to the invoice and populate
it with the information about the preceding invoice.